server/requirements.txt
fastapi
uvicorn
qdrant-client
openai
psycopg2-binary
python-dotenv
pydantic
python-multipart
chunking-library
mdx-parser
passlib[bcrypt]
pyjwt
